[kernel] De-duplicate console code, update BIOS console
